Separate player files for better compartmentalization
